requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attr==0.3.1
attrs==23.1.0
azure-core==1.29.4
azure-storage-blob==12.18.2
bleach==5.0.1
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.7.1
Django==3.2.20
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
expiringdict==1.2.2
google-api-core==2.11.0
google-auth==2.14.1
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.59.0
grpcio-status==1.59.0
htmlmin==0.1.12
idna==3.4
ijson==3.2.3
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==3.2.0
label-studio==1.8.2.post1
label-studio-converter==0.0.54rc0
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
lockfile==0.12.2
lxml==4.9.3
MarkupSafe==2.1.3
nltk==3.6.7
numpy==1.24.3
openapi-codec==1.3.2
ordered-set==4.0.2
packaging==23.2
pandas==2.1.1
Pillow==9.3.0
proto-plus==1.22.3
protobuf==4.24.3
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
pyRFC3339==1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.1
redis==3.5.3
regex==2023.8.8
requests==2.31.0
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
semver==2.13.0
sentry-sdk==1.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
ujson==5.8.0
uritemplate==4.1.1
urllib3==1.26.16
user-agents==2.2.0
webencodings==0.5.1
xmljson==0.2.0
